requirements.txt
psycopg2-binary==2.9.6
cloudinary==1.32.0
colorama==0.4.6
dj-database-url==1.3.0
Django==3.2.18
django-cors-headers==3.14.0
django-environ==0.9.0
django-filter==23.1
django-redis==5.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-spectacular==0.26.2
gunicorn==20.1.0
huggingface-hub==0.13.4
locust==2.15.1
openai==0.27.4
PyJWT==2.6.0
python-dotenv==0.21.1
requests==2.28.2
tokenizers==0.13.3
sentry-sdk==1.21.1
whitenoise==6.4.0
transformers==4.28.1
oauthlib==3.2.2
facebook-sdk==3.1.0
google-api-python-client==2.86.0
oauthlib==3.2.2
python-twitter==3.5
firebase-admin==6.1.0